Separate RPC methods into files
