+++ aquadoggo/src/rpc/methods/mod.rs
@@ -1,0 +1,10 @@
+mod entry_args;
+mod publish_entry;
+
+pub mod error {
+    pub use super::entry_args::EntryArgsError;
+    pub use super::publish_entry::PublishEntryError;
+}
+
+pub use entry_args::get_entry_args;
+pub use publish_entry::publish_entry;
