+++ aquadoggo/src/rpc/methods/publish_entry.rs
@@ -1,17 +1,16 @@
 use std::convert::TryFrom;
 
-use bamboo_rs_core::entry::is_lipmaa_required;
 use p2panda_rs::atomic::{Entry as EntryUnsigned, Message};
 
-use super::request::{EntryArgsRequest, PublishEntryRequest};
-use super::response::{EntryArgsResponse, PublishEntryResponse};
 use crate::db::models::{Entry, Log};
 use crate::db::Pool;
 use crate::errors::Result;
+use crate::rpc::request::PublishEntryRequest;
+use crate::rpc::response::PublishEntryResponse;
 
 #[derive(thiserror::Error, Debug)]
 #[allow(missing_copy_implementations)]
-pub enum ApiError {
+pub enum PublishEntryError {
     #[error("Could not find backlink entry in database")]
     BacklinkMissing,
 
@@ -20,51 +19,6 @@ pub enum ApiError {
 
     #[error("Claimed log_id for schema not the same as in database")]
     InvalidLogId,
-}
-
-/// Implementation of `panda_getEntryArguments` RPC method.
-///
-/// Returns required data (backlink and skiplink entry hashes, last sequence number and the schemas
-/// log_id) to encode a new bamboo entry.
-pub async fn get_entry_args(pool: Pool, params: EntryArgsRequest) -> Result<EntryArgsResponse> {
-    // Determine log_id for author's schema
-    let log_id = Log::find_schema_log_id(&pool, &params.author, &params.schema).await?;
-
-    // Find latest entry in this log
-    let entry_latest = Entry::latest(&pool, &params.author, &log_id).await?;
-
-    match entry_latest {
-        Some(entry_backlink) => {
-            let next_seq_num = entry_backlink.seq_num.clone().next().unwrap();
-
-            // Unwrap as we know that an skiplink exists as soon as previous entry is given:
-            let skiplink_seq_num = next_seq_num.skiplink_seq_num().unwrap();
-
-            // Determine skiplink ("lipmaa"-link) entry in this log, return `None` when no skiplink
-            // is required for the next entry
-            let entry_hash_skiplink = if is_lipmaa_required(next_seq_num.as_i64() as u64) {
-                let entry = Entry::at_seq_num(&pool, &params.author, &log_id, &skiplink_seq_num)
-                    .await?
-                    .unwrap();
-                Some(entry.entry_hash)
-            } else {
-                None
-            };
-
-            Ok(EntryArgsResponse {
-                entry_hash_backlink: Some(entry_backlink.entry_hash),
-                entry_hash_skiplink,
-                last_seq_num: Some(entry_backlink.seq_num),
-                log_id,
-            })
-        }
-        None => Ok(EntryArgsResponse {
-            entry_hash_backlink: None,
-            entry_hash_skiplink: None,
-            last_seq_num: None,
-            log_id,
-        }),
-    }
 }
 
 /// Implementation of `panda_publishEntry` RPC method.
@@ -87,7 +41,7 @@ pub async fn publish_entry(
 
     // Check if log_id is the same as the previously claimed one (when given)
     if schema_log_id.is_some() && schema_log_id.as_ref() != Some(entry.log_id()) {
-        Err(ApiError::InvalidLogId)?;
+        Err(PublishEntryError::InvalidLogId)?;
     }
 
     // Get related bamboo backlink and skiplink entries
@@ -105,7 +59,7 @@ pub async fn publish_entry(
                     .expect("Backlink entry with invalid hex-encoding detected in database"),
             )
         })
-        .ok_or(ApiError::BacklinkMissing)
+        .ok_or(PublishEntryError::BacklinkMissing)
     } else {
         Ok(None)
     }?;
@@ -124,7 +78,7 @@ pub async fn publish_entry(
                     .expect("Skiplink entry with invalid hex-encoding detected in database"),
             )
         })
-        .ok_or(ApiError::SkiplinkMissing)
+        .ok_or(PublishEntryError::SkiplinkMissing)
     } else {
         Ok(None)
     }?;
@@ -170,11 +124,7 @@ mod tests {
     use p2panda_rs::key_pair::KeyPair;
 
     use crate::rpc::ApiService;
-    use crate::test_helpers::{
-        initialize_db, random_entry_hash, rpc_error, rpc_request, rpc_response,
-    };
-
-    const TEST_AUTHOR: &str = "8b52ae153142288402382fd6d9619e018978e015e6bc372b1b0c7bd40c6a240a";
+    use crate::test_helpers::{initialize_db, rpc_error, rpc_request, rpc_response};
 
     // Helper method to create encoded entries and messages
     fn create_test_entry(
@@ -200,35 +150,6 @@ mod tests {
         let entry_encoded = EntrySigned::try_from((&entry, key_pair)).unwrap();
 
         (entry_encoded, message_encoded)
-    }
-
-    #[async_std::test]
-    async fn get_entry_arguments() {
-        let pool = initialize_db().await;
-        let io = ApiService::io_handler(pool);
-
-        let request = rpc_request(
-            "panda_getEntryArguments",
-            &format!(
-                r#"{{
-                    "author": "{}",
-                    "schema": "{}"
-                }}"#,
-                TEST_AUTHOR,
-                random_entry_hash(),
-            ),
-        );
-
-        let response = rpc_response(
-            r#"{
-                "entryHashBacklink": null,
-                "entryHashSkiplink": null,
-                "lastSeqNum": null,
-                "logId": 1
-            }"#,
-        );
-
-        assert_eq!(io.handle_request_sync(&request), Some(response));
     }
 
     #[async_std::test]
